+++ index.js
@@ -4,6 +4,7 @@ module.exports = {
   rules: {
     'consistent-compose': require('./rules/consistent-compose'),
     'consistent-name': require('./rules/consistent-name'),
+    'no-argumentless-calls': require('./rules/no-argumentless-calls'),
     'no-chain': require('./rules/no-chain'),
     'no-extraneous-function-wrapping': require('./rules/no-extraneous-function-wrapping'),
     'no-single-composition': require('./rules/no-single-composition'),
@@ -23,6 +24,7 @@ module.exports = {
       rules: {
         'lodash-fp/consistent-compose': 'off',
         'lodash-fp/consistent-name': ['error', '_'],
+        'lodash-fp/no-argumentless-calls': 'error',
         'lodash-fp/no-chain': 'error',
         'lodash-fp/no-extraneous-function-wrapping': 'error',
         'lodash-fp/no-single-composition': 'error',